# src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: HalaFood API
  version: 1.0.0
  description: HalaFood backend API documentation.
servers:
  - url: 'http://localhost:3000/api'
    description: API base path
tags:
  - name: Categories
    description: Main categories and subcategories
  - name: Products
    description: Product management
paths:
  /main-categories:
    post:
      tags:
        - Categories
      summary: Create a main category
      operationId: createMainCategory
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the main category
                image:
                  type: string
                  format: binary
                  description: Image file for the main category
              required:
                - name
                - image
      responses:
        '201':
          description: Main category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MainCategoryResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Categories
      summary: List main categories with their subcategories (paginated)
      operationId: getMainCategoriesWithSubcategories
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of items per page
      responses:
        '200':
          description: Paginated list of main categories with subcategories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMainCategoriesResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/main-categories/{id}':
    put:
      tags:
        - Categories
      summary: Update a main category
      operationId: updateMainCategory
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the main category
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: New name of the main category
                image:
                  type: string
                  format: binary
                  description: New image file for the main category
      responses:
        '200':
          description: Main category updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MainCategoryResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Main category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Categories
      summary: Delete a main category
      operationId: deleteMainCategory
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the main category
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Main category deleted successfully
        '404':
          description: Main category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/main-categories/{mainCategoryId}/subcategories':
    post:
      tags:
        - Categories
      summary: Create a subcategory for a main category
      operationId: createSubcategory
      parameters:
        - in: path
          name: mainCategoryId
          required: true
          description: ID of the main category
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubcategoryRequest'
      responses:
        '201':
          description: Subcategory created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubcategoryResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Main category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/main-categories/{mainCategoryId}/subcategories/{subcategoryId}':
    put:
      tags:
        - Categories
      summary: Update a subcategory
      operationId: updateSubcategory
      parameters:
        - in: path
          name: mainCategoryId
          required: true
          description: ID of the main category
          schema:
            type: string
            format: uuid
        - in: path
          name: subcategoryId
          required: true
          description: ID of the subcategory
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubcategoryRequest'
      responses:
        '200':
          description: Subcategory updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubcategoryResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Main category or subcategory not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/subcategories/{subcategoryId}':
    delete:
      tags:
        - Categories
      summary: Delete a subcategory
      operationId: deleteSubcategory
      parameters:
        - in: path
          name: subcategoryId
          required: true
          description: ID of the subcategory
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Subcategory deleted successfully
        '404':
          description: Subcategory not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products:
    post:
      tags:
        - Products
      summary: Create a new product
      operationId: createProduct
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        msg:
                          type: string
                        param:
                          type: string
                        location:
                          type: string
                required:
                  - status
                  - message
        '404':
          description: Subcategory not found
          content:
            application/json:
              schema:
                $ref: '#/paths/~1products/get/responses/500/content/application~1json/schema'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/paths/~1products/get/responses/500/content/application~1json/schema'
    get:
      tags:
        - Products
      summary: List all products
      operationId: getAllProducts
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
                required:
                  - status
                  - message
  '/products/{id}':
    get:
      tags:
        - Products
      summary: Get product by ID
      operationId: getProductById
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the product
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/paths/~1products/get/responses/500/content/application~1json/schema'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/paths/~1products/get/responses/500/content/application~1json/schema'
    put:
      tags:
        - Products
      summary: Update a product
      operationId: updateProduct
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the product
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/paths/~1products/post/responses/400/content/application~1json/schema'
        '404':
          description: Product or Subcategory not found
          content:
            application/json:
              schema:
                $ref: '#/paths/~1products/get/responses/500/content/application~1json/schema'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/paths/~1products/get/responses/500/content/application~1json/schema'
    delete:
      tags:
        - Products
      summary: Delete a product
      operationId: deleteProduct
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the product
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Product deleted successfully
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/paths/~1products/get/responses/500/content/application~1json/schema'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/paths/~1products/get/responses/500/content/application~1json/schema'
  /products/visible:
    get:
      tags:
        - Products
      summary: Get visible products with pagination
      operationId: getVisibleProducts
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of items per page
      responses:
        '200':
          description: Paginated list of visible products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductsResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/paths/~1products/get/responses/500/content/application~1json/schema'
  /subcategories:
    get:
      tags:
        - Categories
      summary: Get subcategories with pagination
      operationId: getSubcategories
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of items per page
      responses:
        '200':
          description: Paginated list of subcategories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSubcategoriesResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    MainCategory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        imageUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - imageUrl
        - createdAt
    Subcategory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mainCategoryId:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - mainCategoryId
        - name
        - createdAt
    MainCategoryWithSubcategories:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        imageUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        subcategories:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              createdAt:
                type: string
                format: date-time
            required:
              - id
              - name
              - createdAt
      required:
        - id
        - name
        - imageUrl
        - createdAt
        - subcategories
    CreateSubcategoryRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: failed
        message:
          type: string
      required:
        - status
        - message
    ValidationErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: failed
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              msg:
                type: string
              param:
                type: string
              location:
                type: string
      required:
        - status
        - message
    MainCategoryResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/MainCategory'
      required:
        - status
        - data
    PaginatedMainCategoriesResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: object
          properties:
            categories:
              type: array
              items:
                $ref: '#/components/schemas/MainCategoryWithSubcategories'
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
            totalPages:
              type: integer
      required:
        - status
        - data
    SubcategoryResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/Subcategory'
      required:
        - status
        - data
    PaginatedSubcategoriesResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: object
          properties:
            subcategories:
              type: array
              items:
                $ref: '#/components/schemas/Subcategory'
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
            totalPages:
              type: integer
      required:
        - status
        - data
    ProductStatus:
      type: string
      enum:
        - visible
        - hidden
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        subcategoryName:
          type: string
        manufacturingLocation:
          type: string
        availableQuantity:
          type: integer
        price:
          type: number
        size:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - subcategoryName
        - manufacturingLocation
        - availableQuantity
        - price
        - size
        - status
        - createdAt
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        subcategoryName:
          type: string
        manufacturingLocation:
          type: string
        availableQuantity:
          type: integer
        price:
          type: number
        size:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
      required:
        - name
        - subcategoryName
        - manufacturingLocation
        - availableQuantity
        - price
        - size
        - status
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        subcategoryName:
          type: string
        manufacturingLocation:
          type: string
        availableQuantity:
          type: integer
        price:
          type: number
        size:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
    ProductResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/Product'
      required:
        - status
        - data
    ProductsListResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'
      required:
        - status
        - data
    PaginatedProductsResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: object
          properties:
            products:
              type: array
              items:
                $ref: '#/components/schemas/Product'
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
            totalPages:
              type: integer